Update tests to reflect new notifications
Unfortunately, the change to the notifications in the previous commit caused
a redesign and refactor of the original testing approach.

Basically, this boils down to testing individual methods - instead of just
testing the Slack POST. While this isn't as simple, it does mean the tests
now exercise the methods more throughly and is more consistent with general
testing approaches.